[2] Create a load test items
`TestItem`s are a managed type in the VS Code API, and created by calling
`controller.createTestItem`. Here, we use that API to create the items.
Children are also handled by the `children` property, which we'll use
to build the nested suite and add the top-level items. Note that, unlike
the Test Explorer UI, there's no "root" element.

We create items when the `resolveHandler` is called. You can load all
tests when this is called, or you can load them gradually as files are
opened. Check out `TestItem.canResolveChildren` for more info about this.

We won't demo it here, but one advantage of being managed types is that
any updates to TestItems' children or properties are automatically
reflected in the UI. No events needed!