[3] Update test running code
VS Code uses the notion of "run profiles" to represent different ways
tests can be run (or debugged). Each profile has a handler, which it'll
call when the user wants to run tests.

When it's called, we'll create a TestRun in response to the TestRunRequest,
and start invoking the test items. If `request.include` is undefined,
that means we should run all tests. Then, we repurpose the existing
"runNode" function to use the native API, updating the state of children
in the test run.